+++ fltk/tests/app_handle.rs
@@ -1,33 +1,34 @@
 use fltk::{app, button::*, frame::*, window::*};
 
 #[test]
-fn messages() {
+fn app_handle() {
     let app = app::App::default();
     let mut wind = Window::default()
         .with_size(400, 300)
         .center_screen()
-        .with_label("messages");
+        .with_label("handle");
     let mut frame = Frame::new(0, 0, 400, 200, "");
     let mut but = Button::new(160, 210, 80, 40, "Click me!");
 
     wind.show();
 
-    let (s, r) = app::channel::<i32>();
-
     but.set_callback(move || {
         std::thread::spawn(move || {
-            for i in 0..1000 {
+            for i in 31..1000 {
                 app::sleep(0.010);
-                s.send(i);
+                let ret = app::handle_main(i);
+                if let Ok(ret) = ret {
+                    println!("Handled? {}", ret);
+
+                }
             }
         });
     });
 
-    while app.wait() {
-        let msg = r.recv();
-        match msg {
-            Some(val) => frame.set_label(format!("Hello {}", val).as_str()),
-            None => (),
-        }
-    }
+    frame.handle2(|f, ev| if ev as i32 > 30 {
+        f.set_label(&format!("{:?}", ev));
+        true
+    } else { false });
+
+    app.run().unwrap();
 }