+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 
+## [0.15.3] - 2021-02-24
+### Changes
+- Mark app::handle and handle_main as safe.
+- Add methods to get and set Dial angles.
+- Add app::remove_idle, has_idle, remove_timeout.
+
 ## [0.15.2] - 2021-02-22
 ### Changes
 - [BREAKING] RgbImage::new() and draw::draw_image() take a ColorDepth enum instead of an int.